refactor(apm): inject shared Datadog tracer into database
