+++ packages/core/src/db/client.ts
@@ -16,6 +16,7 @@ import { drizzle as drizzlePostgres } from 'drizzle-orm/postgres-js';
 import postgres from 'postgres';
 import { loadConfigSync } from '../config/config-manager';
 import type { AgorConfig } from '../config/types';
+import type { DatadogTracer } from '../tracing/datadog';
 import {
   coordinateInMemorySQLiteClient,
   coordinateInMemorySQLiteDatabase,
@@ -204,7 +205,16 @@ async function configureSQLitePragmas(client: Client): Promise<void> {
  * });
  * ```
  */
-export function createDatabase(config: DbConfig): RawDatabase {
+/**
+ * Optional runtime hooks for database creation. `tracer` is the resolved
+ * dd-trace tracer, injected by the daemon so Postgres queries emit APM spans;
+ * `@agor/core` never resolves dd-trace itself (see postgres-tracing.ts).
+ */
+export interface CreateDatabaseOptions {
+  tracer?: DatadogTracer | null;
+}
+
+export function createDatabase(config: DbConfig, options: CreateDatabaseOptions = {}): RawDatabase {
   // Auto-detect dialect from URL if not explicitly set
   let dialect = config.dialect;
 
@@ -223,7 +233,7 @@ export function createDatabase(config: DbConfig): RawDatabase {
   }
 
   if (dialect === 'postgresql') {
-    return createPostgresDatabase(config) as unknown as RawDatabase;
+    return createPostgresDatabase(config, options.tracer) as unknown as RawDatabase;
   }
 
   return createSQLiteDatabase(config) as unknown as RawDatabase;
@@ -232,7 +242,10 @@ export function createDatabase(config: DbConfig): RawDatabase {
 /**
  * Create PostgreSQL database client
  */
-function createPostgresDatabase(config: DbConfig): PostgresJsDatabase<typeof postgresSchema> {
+function createPostgresDatabase(
+  config: DbConfig,
+  tracer?: DatadogTracer | null
+): PostgresJsDatabase<typeof postgresSchema> {
   try {
     // Build options without ssl key by default — postgres.js treats an explicitly-present
     // `ssl: undefined` differently from an absent key. When the key is absent, postgres.js
@@ -277,9 +290,9 @@ function createPostgresDatabase(config: DbConfig): PostgresJsDatabase<typeof pos
     const db = drizzlePostgres(sql, { schema: postgresSchema });
     // Emit `postgres.query` APM spans for every query. dd-trace has no
     // postgres.js plugin, so without this the daemon's DB layer is invisible in
-    // Datadog. Best-effort and additive — a no-op when dd-trace isn't loaded and
-    // it can never break a query (see postgres-tracing.ts).
-    instrumentDrizzlePostgresForTracing(db);
+    // Datadog. Best-effort and additive — a no-op unless the daemon injected a
+    // tracer, and it can never break a query (see postgres-tracing.ts).
+    instrumentDrizzlePostgresForTracing(db, { tracer });
     return db;
   } catch (error) {
     throw new DatabaseConnectionError(
@@ -316,7 +329,10 @@ function createSQLiteDatabase(config: DbConfig): LibSQLDatabase<typeof sqliteSch
  * @param config Database configuration
  * @returns Promise resolving to Drizzle database instance
  */
-export async function createDatabaseAsync(config: DbConfig): Promise<RawDatabase> {
+export async function createDatabaseAsync(
+  config: DbConfig,
+  options: CreateDatabaseOptions = {}
+): Promise<RawDatabase> {
   // Determine dialect: use config.dialect, then auto-detect from URL, then fallback to env/default
   let dialect = config.dialect;
   if (!dialect && config.url) {
@@ -328,7 +344,7 @@ export async function createDatabaseAsync(config: DbConfig): Promise<RawDatabase
 
   if (dialect === 'postgresql') {
     // PostgreSQL doesn't need pragma configuration
-    return createPostgresDatabase(config) as unknown as RawDatabase;
+    return createPostgresDatabase(config, options.tracer) as unknown as RawDatabase;
   }
 
   // SQLite: Wait for pragmas to be configured
